test: move integration tests under integration/ folder
