+++ tsconfig.lib.cjs.json
@@ -2,12 +2,15 @@
   "extends": "./tsconfig.json",
   "compilerOptions": {
     "composite": true,
+    "rootDir": "src",
     "module": "CommonJS",
     "target": "ES2015",
     "outDir": "dist/cjs",
   },
-  "include": ["src/**/*.ts"],
+  "include": [
+    "src/**/*.ts"
+  ],
   "exclude": [
-    "src/**/*.test.ts"
+    "**/*.test.ts"
   ]
 }
