+++ tsconfig.test.json
@@ -2,6 +2,7 @@
   "extends": "./tsconfig.json",
   "compilerOptions": {
     "composite": true,
+    "rootDir": ".",
     "module": "ESNext",
     "target": "ESNext",
     "moduleResolution": "node",
